+++ packages/liveblocks-react/test-d/no-augmentation.test-d.tsx
@@ -1,7 +1,13 @@
 import React from "react";
 React; // To silence tsd warning
 
-import type { BaseMetadata, Json, Lson } from "@liveblocks/client";
+import {
+  BaseMetadata,
+  Json,
+  LiveList,
+  LiveObject,
+  Lson,
+} from "@liveblocks/client";
 import * as classic from "@liveblocks/react";
 import * as suspense from "@liveblocks/react/suspense";
 import { expectAssignable, expectError, expectType } from "tsd";
@@ -56,6 +62,94 @@ import { expectAssignable, expectError, expectType } from "tsd";
     authEndpoint="/api/auth"
     resolveUsers={async () => [{ name: "Vincent", age: 42 }]}
   />;
+}
+
+// RoomProvider
+{
+  const RoomProvider = classic.RoomProvider;
+
+  // Missing mandatory room ID is an error
+  expectError(
+    <RoomProvider /* no room id */>
+      <div />
+    </RoomProvider>
+  );
+
+  <RoomProvider id="my-room">
+    <div />
+  </RoomProvider>;
+
+  <RoomProvider
+    id="my-room"
+    initialPresence={{ anything: ["is", "fine", "here"] }}
+  >
+    <div />
+  </RoomProvider>;
+
+  <RoomProvider
+    id="my-room"
+    initialStorage={{
+      foo: new LiveList([]),
+      bar: new LiveObject(),
+    }}
+  >
+    <div />
+  </RoomProvider>;
+
+  <RoomProvider
+    id="my-room"
+    initialPresence={{ anything: ["is", "fine", "here"] }}
+    initialStorage={{
+      foo: new LiveList([]),
+      bar: new LiveObject(),
+    }}
+  >
+    <div />
+  </RoomProvider>;
+}
+
+// RoomProvider (suspense)
+{
+  const RoomProvider = suspense.RoomProvider;
+
+  // Missing mandatory room ID is an error
+  expectError(
+    <RoomProvider /* no room id */>
+      <div />
+    </RoomProvider>
+  );
+
+  <RoomProvider id="my-room">
+    <div />
+  </RoomProvider>;
+
+  <RoomProvider
+    id="my-room"
+    initialPresence={{ anything: ["is", "fine", "here"] }}
+  >
+    <div />
+  </RoomProvider>;
+
+  <RoomProvider
+    id="my-room"
+    initialStorage={{
+      foo: new LiveList([]),
+      bar: new LiveObject(),
+    }}
+  >
+    <div />
+  </RoomProvider>;
+
+  <RoomProvider
+    id="my-room"
+    initialPresence={{ anything: ["is", "fine", "here"] }}
+    initialStorage={{
+      foo: new LiveList([]),
+      bar: new LiveObject(),
+    }}
+  >
+    <div />
+  </RoomProvider>;
 }
 
 // ---------------------------------------------------------